refactor: Change Incomes command to a specific one representing the create function
PS: This was done to decouple the commands. Instead of just one class
containing all methods, we created different ones.

PS2: Changes class command to function command. This was mainly
because the 'this' variable from the class was undefined so we
could not declare any variables using it. We will probably
refactor it in the nearly future.